feat: Script 로드 성공 여부에 따라 redux state 업데이트
- Next Script Strategy 변경: beforeInteractive -> afterInteractive
